secStudent/pip_requirements.txt
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
easydict==1.9
idna==2.10
imageio==2.9.0
imutils==0.5.4
kiwisolver==1.3.1
llvmlite==0.36.0
matplotlib==3.4.2
networkx==2.5.1
numba==0.53.1
numpy @ file:///D:/bld/numpy_1621506450533/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.5.2.52
pandas==1.2.4
Pillow @ file:///C:/ci/pillow_1617386341487/work
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
tifffile==2021.4.8
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1
tqdm @ file:///tmp/build/80754af9/tqdm_1615925068909/work
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1602702424206/work
urllib3==1.26.5
wincertstore==0.2
